requirements.txt
absl-py==2.3.0
accelerate==1.7.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.7
aiosignal==1.3.2
alembic==1.16.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
arrow==1.3.0
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
av==14.4.0
babel==2.17.0
backoff==1.11.1
beautifulsoup4==4.13.4
binaryornot==0.4.4
black==25.1.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cloudpickle==3.1.1
codecarbon==3.0.2
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.2
cookiecutter==1.7.3
cryptography==45.0.3
csvw==3.5.1
cycler==0.12.1
datasets==3.6.0
decorator==5.2.1
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@65184fc057d4fab080a98564f6b60fae0b94edc4
dill==0.3.4
dlinfo==2.0.0
docopt==0.6.2
evaluate==0.4.3
exceptiongroup==1.3.0
execnet==2.1.1
faiss-gpu==1.7.2
fastapi==0.115.12
ffmpy==0.6.0
fief-client==0.20.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.1
frozenlist==1.6.2
fsspec==2025.3.0
ftfy==6.3.1
fugashi==1.4.3
fvcore==0.1.5.post20221221
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.18
gradio==5.33.2
gradio_client==1.10.3
greenlet==3.2.2
groovy==0.1.2
grpcio==1.72.1
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httpx==0.27.2
huggingface-hub==0.32.4
humanfriendly==10.0
hydra-core==1.3.2
hypothesis==6.135.0
idna==3.10
iniconfig==2.1.0
iopath==0.1.9
ipadic==1.0.0
isodate==0.7.2
isort==6.0.1
itsdangerous==2.0.1
Jinja2==3.1.6
jinja2-time==0.2.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jwcrypto==1.5.6
kenlm==0.3.0
kiwisolver==1.4.8
langchain-core==0.3.65
langgraph==0.4.8
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.3.45
language-tags==1.2.0
lazy_loader==0.4
libcst==1.8.0
librosa==0.11.0
llvmlite==0.44.0
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.4
multiprocess==0.70.12.2
mypy_extensions==1.1.0
networkx==3.4.2
nltk==3.8.1
num2words==0.5.14
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
onnx==1.18.0
onnxruntime==1.22.0
onnxruntime-tools==1.7.0
open-clip-torch==2.23.0
optuna==4.3.0
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pandas==2.2.3
parameterized==0.9.0
pathspec==0.12.1
phonemizer==3.3.0
pillow==11.2.1
plac==1.4.5
platformdirs==4.3.8
pluggy==1.6.0
pooch==1.8.2
portalocker==2.0.0
poyo==0.5.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==6.31.1
psutil==7.0.0
py-cpuinfo==9.0.0
py3nvml==0.2.7
pyarrow==20.0.0
pycocotools==2.0.9
pycparser==2.22
pyctcdecode==0.5.0
pydantic==2.11.5
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
pygtrie==2.5.0
pynvml==12.0.0
pyparsing==3.2.3
pypng==0.20220715.0
pytesseract==0.3.13
pytest==8.4.0
pytest-asyncio==1.0.0
pytest-order==1.3.0
pytest-rerunfailures==15.1
pytest-rich==0.2.0
pytest-timeout==2.4.0
pytest-xdist==3.7.0
python-dateutil==2.9.0.post0
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
questionary==2.1.0
RapidFuzz==3.13.0
ray==2.46.0
rdflib==7.1.4
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==1.5.0
rhoknp==1.3.0
rich==14.0.0
rjieba==0.1.13
rouge-score==0.1.2
rpds-py==0.25.1
ruff==0.11.2
sacrebleu==1.5.1
sacremoses==0.1.1
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
segments==2.3.0
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
sigopt==8.8.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
SQLAlchemy==2.0.41
starlette==0.46.2
SudachiDict-core==20250515
SudachiPy==0.6.10
sympy==1.13.1
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.3.0
text-unidecode==1.3
threadpoolctl==3.6.0
timeout-decorator==0.5.0
timm==1.0.11
tokenizers==0.21.1
tomli==2.2.1
tomlkit==0.13.3
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
-e git+https://github.com/huggingface/transformers@78d771c3c21922642fc9546ccb973cc7a182ab34#egg=transformers
triton==3.2.0
typer==0.16.0
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
unidic==1.1.0
unidic-lite==1.0.8
uritemplate==4.2.0
urllib3==1.26.20
uvicorn==0.34.3
wasabi==0.10.1
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
xmltodict==0.14.2
xxhash==3.5.0
yacs==0.1.8
yarl==1.20.0
yaspin==3.1.0
zstandard==0.23.0
